strip out as many of the intermediate objects as I can
and generate the JSON directly. I still have to assemble it piecemeal,
because of the way pickle can reference past objects, but at least
I don't have a lot of complex types lying around. Apart from Tuple,
everything is either a string or a strings.Builder. (Tuple is used
to build function arguments, and I need to be able to unpack it)