remove another 5% of allocations by not even copying a string
out of the initial []byte, and not putting " around it until
it's JSON time.

As a consequence, we don't un-escape when String() is called.
So far that's not a problem b/c none of the class and module
names are unicode. But if they were then FindClass would have
to accept JSON-escaped names.