my use cases all take a []byte as input, and once we know that,
we can be more efficient not only because we don't have to perform
IO through the Read and ReadByte interface calls, but also we don't
have to allocate any read buffers at all. Instead we return the
slice of the input which is being read.

This improves allocs by ~5%, allocated bytes by ~50%, and CPU time
by ~15% to ~20%.