handle strings which need escaping differently
instead of escaping them when the string is parsed, escape them while
writing the string to the output. This avoids an alloc (or more) for
the buffer were holding the escaped string in, and makes String() string
correct even for escaped strings.

Note that scanning the string at NewString() time to see if it needs
escaping is faster than scanning it at JSON() time. I bet that's because
at NewString() time the string has been prefetched by the cache,
while it's colder when we are calling JSON().